Fix walking bug due to int not being cast into Integer
r5070
2019-10-04

## CHANGE
1. Temporarily disable the 'restart menu' option in beryx
   console until it gets fixed in future.
   - Comment out restartItem in configureMainMenu() in
     MarsTerminal.
2. Reduce EVA_SUIT_VALUE four-fold from 20 to 5 in
   GoodsManager so that it's VP is smaller than that of LUV.
3. Add checking if a person is hungry in canExploreSite()
   in ExploreSite.
4. Simplify electAnOperator() in Airlock.

## FIX
1. The weight of an EVA suit should be the sum of all parts.
   - Adjust the weights of the parts.
2. When setting the value of operatorID in Airlock, cast it
   to Integer.
   - Compare operatorID to Integer.valueOf(-1)
3. Enforce casting id into Integer in Equipment, Unit,
   MBTIPersonality, TraitManager, ScientificManager,
   ScientificStudy
4. Correct Airlock class issues in that lookUpPerson in
   UnitManager cannot be read. Reasons still unknown.
   - Create a static lookUpPerson in Airlock to track the
     id of a person. This alleviates the constant calling
	 of lookUpPerson in UnitManager.

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/equipment/Equipment.java
@@ -47,7 +47,7 @@ public abstract class Equipment extends Unit implements Indoor, Salvagable {
 	/** Unique identifier for this equipment. */
 	private int identifier;
 	/** The identifier for the last owner of this equipment. */
-	private int lastOwner;
+	private Integer lastOwner;
 	/** The equipment type. */
 	private String type;
 	/** The SalvageInfo instatnce. */	
@@ -90,7 +90,7 @@ public abstract class Equipment extends Unit implements Indoor, Salvagable {
 		isSalvaged = false;
 		salvageInfo = null;
 		
-		lastOwner = -1;
+		lastOwner = Integer.valueOf(-1);
 		
 		this.identifier = getNextIdentifier();
 		// Add this equipment to the equipment lookup map	
@@ -118,7 +118,7 @@ public abstract class Equipment extends Unit implements Indoor, Salvagable {
 		Collection<Person> people = new ConcurrentLinkedQueue<Person>();
 
 		Person owner = null;
-		if (lastOwner != -1) {// && lastOwner instanceof Person) {
+		if (lastOwner != Integer.valueOf(-1)) {// && lastOwner instanceof Person) {
 			owner = unitManager.getPersonByID(lastOwner);
 			people.add(owner);
 		}
@@ -283,11 +283,11 @@ public abstract class Equipment extends Unit implements Indoor, Salvagable {
 	 */
 	public void setLastOwner(Unit unit) {
 		if (unit != null) {
-			 if (unit.getIdentifier() != lastOwner)
-				 lastOwner = unit.getIdentifier();
+			 if ((Integer) unit.getIdentifier() != lastOwner)
+				 lastOwner = (Integer) unit.getIdentifier();
 		}	
 		else
-			lastOwner = Unit.UNKNOWN_ID;
+			lastOwner = (Integer) Unit.UNKNOWN_ID;
 	}
 
 	/**
